[cegar-ocp] remove compute_num_states_in_task() function

+++ src/search/utils/math.cc
@@ -3,8 +3,7 @@
 #include <cassert>
 
 namespace utils {
-bool is_product_within_limit(
-    long long factor1, long long factor2, long long limit) {
+bool is_product_within_limit(int factor1, int factor2, int limit) {
     assert(factor1 >= 0);
     assert(factor2 >= 0);
     assert(limit >= 0);
